Assessment/Conversational_AI/requirements.txt
streamlit
google-generativeai
langchain
langchain-google-genai
autogen
python-dotenv